requirements.txt
sentence-transformers>=2.7.0
bert-score>=0.3.13
transformers>=4.40.0
torch>=2.2.0
spacy>=3.7.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
matplotlib>=3.8.0
seaborn>=0.13.0
anthropic>=0.28.0
huggingface_hub>=0.23.0
trafilatura>=1.6.0
requests>=2.31.0
python-dotenv>=1.0.0
rank-bm25>=0.2.2
